+++ taf/utils.py
@@ -22,8 +22,7 @@ from json import JSONDecoder
 import taf.settings
 from taf.exceptions import PINMissmatchError
 from taf.log import taf_logger
-from typing import List, Optional, Tuple, Dict, Any
-import securesystemslib
+from typing import List, Optional, Tuple, Dict
 from securesystemslib.hash import digest_fileobject
 from securesystemslib.storage import FilesystemBackend, StorageBackendInterface
 
@@ -318,34 +317,24 @@ def get_file_details(
 
     # Making sure that the format of 'filepath' is a path string.
     if not isinstance(filepath, str) or not filepath:
-        raise securesystemslib.exceptions.FormatError(
-            "The filepath must be a non-empty string."
-        )
+        raise ValueError("The filepath must be a non-empty string.")
 
     if not isinstance(hash_algorithms, list):
-        raise securesystemslib.exceptions.FormatError(
-            "The hash_algorithms must be a list."
-        )
+        raise ValueError("The hash_algorithms must be a list.")
     for algo in hash_algorithms:
-        if algo not in hash_algorithms:  # Add any other valid algorithms as needed
-            raise securesystemslib.exceptions.FormatError(
-                f"Invalid hash algorithm: {algo}"
-            )
+        if algo not in ["sha256", "sha512"]:  # Add any other valid algorithms as needed
+            raise ValueError(f"Invalid hash algorithm: {algo}")
 
     if storage_backend is None:
         storage_backend = FilesystemBackend()
 
     # Getting the file length
     if not os.path.isabs(filepath):
-        raise securesystemslib.exceptions.FormatError(
-            "The 'filepath' must be an absolute path"
-        )
+        raise ValueError("The 'filepath' must be an absolute path")
 
     # Check if the file exists and get its size
     if not os.path.isfile(filepath):
-        raise securesystemslib.exceptions.StorageError(
-            f"The file at '{filepath}' cannot be opened or found"
-        )
+        raise FileNotFoundError(f"The file at '{filepath}' cannot be opened or found")
 
     file_length = os.path.getsize(filepath)
 
@@ -360,33 +349,6 @@ def get_file_details(
             )  # Reset file object position after reading for the next hash
 
     return file_length, file_hashes
-
-
-def load_json_file(
-    filepath: str, storage_backend: Optional[StorageBackendInterface] = None
-) -> Any:
-    if not isinstance(filepath, str) or not filepath:
-        raise securesystemslib.exceptions.FormatError(
-            "The filepath must be a non-empty string."
-        )
-
-    if storage_backend is None:
-        storage_backend = FilesystemBackend()
-
-    deserialized_object = None
-    with storage_backend.get(filepath) as file_obj:
-        raw_data = file_obj.read().decode("utf-8")
-
-        try:
-            deserialized_object = json.loads(raw_data)
-
-        except (ValueError, TypeError):
-            raise securesystemslib.exceptions.Error(
-                "Cannot deserialize to a" " Python object: " + filepath
-            )
-
-        else:
-            return deserialized_object
 
 
 class timed_run: